Refactor: Relocate DTOs and update serialization
- Moved `BaseResponseDto` and `PagedDataDto` to the `shared.dto` package.
- Added `@Serializable` and `@SerialName` annotations to `BaseResponseDto` and `PagedDataDto` for Kotlinx serialization.
- Updated import statements across various files to reflect the new DTO locations.
- Adjusted null handling in `ContactMapper.kt` for `PagedDataDto` properties.
- Corrected the Kotlin serialization plugin version in `gradle/libs.versions.toml`.
