+++ core_chat_data/src/commonMain/kotlin/net/thechance/mena/core_chat/data/contacts/ContactMapper.kt
@@ -2,7 +2,7 @@ package net.thechance.mena.core_chat.data.contacts
 
 import net.thechance.mena.core_chat.data.contacts.source.remote.dto.ContactCreationRequestDto
 import net.thechance.mena.core_chat.data.contacts.source.remote.dto.ContactDto
-import net.thechance.mena.core_chat.data.contacts.source.remote.dto.PagedDataDto
+import net.thechance.mena.core_chat.data.shared.dto.PagedDataDto
 import net.thechance.mena.core_chat.domain.entity.Contact
 import net.thechance.mena.core_chat.domain.exception.GetUserContactsException
 import net.thechance.mena.core_chat.domain.model.PagedData
@@ -11,8 +11,8 @@ fun PagedDataDto<ContactDto>?.toPagedListOfContacts(): PagedData<Contact> {
     val pagedData = this ?: throw GetUserContactsException("Response body is null")
     return PagedData(
         data = pagedData.data.toListOfContact(),
-        totalItems = pagedData.totalItems,
-        isLastPage = pagedData.pageNumber >= pagedData.totalPages
+        totalItems = pagedData.totalItems ?: 0,
+        isLastPage = (pagedData.pageNumber ?: 0) >= (pagedData.totalPages ?: 0)
     )
 }
