--- core_chat_data/src/commonMain/kotlin/net/thechance/mena/core_chat/data/contacts/source/remote/BaseResponseDto.kt
@@ -1,8 +1,0 @@
-package net.thechance.mena.core_chat.data.contacts.source.remote
-
-open class BaseResponseDto<T>(
-    val body: T? = null,
-    val status: Int,
-    val success: Boolean,
-    val message: String? = null
-)
