--- core_chat_data/src/commonMain/kotlin/net/thechance/mena/core_chat/data/contacts/source/remote/dto/PagedDataDto.kt
@@ -1,9 +1,0 @@
-package net.thechance.mena.core_chat.data.contacts.source.remote.dto
-
-data class PagedDataDto<T>(
-    val data: List<T>,
-    val pageNumber: Int,
-    val pageSize: Int,
-    val totalItems: Int,
-    val totalPages: Int
-)
