+++ core_chat_data/src/commonMain/kotlin/net/thechance/mena/core_chat/data/shared/BaseRepository.kt
@@ -1,6 +1,6 @@
 package net.thechance.mena.core_chat.data.shared
 
-import net.thechance.mena.core_chat.data.contacts.source.remote.BaseResponseDto
+import net.thechance.mena.core_chat.data.shared.dto.BaseResponseDto
 import net.thechance.mena.core_chat.domain.exception.ChatException
 import net.thechance.mena.core_chat.domain.exception.UnknownException
