+++ core_chat_data/src/commonMain/kotlin/net/thechance/mena/core_chat/data/shared/dto/BaseResponseDto.kt
@@ -1,0 +1,16 @@
+package net.thechance.mena.core_chat.data.shared.dto
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+open class BaseResponseDto<T>(
+    @SerialName("body")
+    val body: T? = null,
+    @SerialName("status")
+    val status: Int? = null,
+    @SerialName("success")
+    val success: Boolean = false,
+    @SerialName("message")
+    val message: String? = null
+)
